# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 电力交易智能辅助决策系统 API
  description: |
    配套开发基线 5.5 API 目录总表。统一响应封装 ApiResponse；分页用 PageResult；
    鉴权 JWT Bearer；错误码分段见 README 第 4 条。所有路径省略网关前缀 /api/v1。
  version: 1.0.0
  license:
    name: 项目内部使用
    url: https://api.ptidss.example.com/license
  contact:
    name: PTIDSS 架构组
servers:
  - url: https://api.ptidss.example.com/api/v1
    description: 生产（示例域名）
  - url: https://api-test.ptidss.example.com/api/v1
    description: 测试
...
security:
  - bearerAuth: []
paths:
...
  /admin/users:
    get:
      operationId: getAdminUsers
      summary: 用户分页
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - { $ref: '#/components/parameters/PageParam' }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [active, disabled, locked] } }
      responses:
        '200':
          description: 分页
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties: { data: { $ref: '#/components/schemas/PageResult' } }
        '4XX': { $ref: '#/components/responses/BizError' }
    post:
      operationId: postAdminUsers
      summary: 新建用户
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, roleIds]
              properties:
                username: { type: string }
                password: { type: string, format: password }
                displayName: { type: string }
                roleIds: { type: array, items: { type: integer } }
                regionCodes: { type: array, items: { type: string } }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
    put:
      operationId: putAdminUsers
      summary: 更新用户
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: string, description: 雪花 ID（JSON 序列化为字符串） }
                displayName: { type: string }
                roleIds: { type: array, items: { type: integer } }
                regionCodes: { type: array, items: { type: string } }
                status: { type: string }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
  /admin/users/{id}:
    get:
      operationId: getAdminUsersDetail
      summary: 用户详情
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 用户
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties: { data: { $ref: '#/components/schemas/SysUser' } }
        '4XX': { $ref: '#/components/responses/BizError' }
    delete:
      operationId: deleteAdminUsers
      summary: 删除用户（软删）
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
  /admin/users/{id}/regions:
    get:
      operationId: getAdminUsersRegions
      summary: 用户已授权区域
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 区域编码列表
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties:
                      data: { type: array, items: { type: string } }
        '4XX': { $ref: '#/components/responses/BizError' }
  /admin/users/{id}/password:
    put:
      operationId: putAdminUsersPassword
      summary: 重置密码
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, format: password }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: pageNo
      in: query
      description: 分页参数（pageNo/pageSize，对齐后端 @RequestParam 实现）
      schema:
        type: object
        properties:
          pageNo: { type: integer, minimum: 1, default: 1 }
          pageSize: { type: integer, maximum: 200, default: 20 }
  responses:
    BizError:
      description: 业务错误（错误码分段见 README）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message, data, traceId]
      properties:
        code: { type: integer, description: 0 成功；非 0 见错误码分段表, example: 0 }
        message: { type: string, example: success }
        data: { type: object, nullable: true, description: 业务数据（随接口变化） }
        traceId: { type: string, description: 全链路追踪 ID }
    ErrorResponse:
      type: object
      properties:
        code: { type: integer, example: 14001 }
        message: { type: string }
        traceId: { type: string }
        detail: { type: object, nullable: true, description: 参数错误明细 }
    PageResult:
      type: object
      properties:
        list: { type: array, items: { type: object } }
        pageNo: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...
    SysUser:
      type: object
      properties:
        id: { type: string, description: 雪花 ID（JSON 序列化为字符串） }
        username: { type: string }
        displayName: { type: string }
        status: { type: string }
        roleIds: { type: array, items: { type: integer } }
        regionCodes: { type: array, items: { type: string } }
...